refactor(stacked-component): update StackedComponent name + location to shared
