+++ src/components/PeerInfo.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 
 // Components
 import Card from "components/styles/Card";
-import StackedData from "components/Stacked/StackedComponent";
+import StackedData from "components/shared/StackedData";
 
 // Util
 import { timeAgo } from "utils/util";
